Moved Header to Navbar.js and set it in index.js
