+++ src/index.js
@@ -8,6 +8,7 @@ import theme from './theme';
 import App from './App';
 import Search from './Search';
 import NotFound from './NotFound';
+import Navbar from './Navbar';
 
 // eslint-disable-next-line react/jsx-filename-extension
 ReactDOM.render(
@@ -15,6 +16,7 @@ ReactDOM.render(
     <BrowserRouter>
       <Normalize />
       <GlobalStyle />
+      <Navbar />
       <Routes>
         <Route path="/" element={<App />} />
         <Route path="search" element={<Search />} />
